Move logic to CommentFlagProcessor
