+++ package.json
@@ -112,6 +112,7 @@
         "@xml-tools/parser": "^1.0.7",
         "array-flat-polyfill": "^1.0.1",
         "chalk": "^2.4.2",
+        "chevrotain": "^7.0.1",
         "chokidar": "^3.0.2",
         "clear": "^0.1.0",
         "cross-platform-clear-console": "^2.3.0",
